\name{levelPValues}
\alias{levelPValues}
\title{Level P-values}
\description{
Function to level out a P-value spectrum generated by the Polyfit extension of DESeq by fitting a quadratic function to the right hand portion of the spectrum, produce 'corrected' p-values and q-values using an adapted version of the Storey-Tibsharini procedure}
\usage{
levelPValues(oldPvals, plot = FALSE)
}
\arguments{
  \item{oldPvals}{
an array of p-values produced by the Polyfit replacement of the DESeq function pfNbinomTest() or the Plyfit replacement of the edgeR function pfExactTest()}
  \item{plot}{
TRUE to plot original and corrected pvalue spectra; FALSE not to plot}
}
\details{
\code{levelPValues} should only be used with P-values generated by the Polyfit function \code{\link{pfNbinomTest}}, and not with P-values generated by \code{\link{nbinomTest}}.
}
\value{
List containing
\item{pi0estimate}{an estimate of the proportion of genes not differentially expressed}
\item{lambdaOptimal}{the point in the p-value spectrum past which a quadratic is fitted}
\item{pValueCorr}{p-values calculated from the levelled spectrum}
\item{qValueCorr}{q-values calculated from the levelled spectrum}
\item{qValueCorrBH}{q-values calculated from pValueCorr using Benjamini-Hochberg}
}
\references{
Burden, C.J., Qureshi, S. and Wilson, S.R. (2014). \emph{Error estimates for the 
analysis of differential expression from RNA-seq count data}, PeerJ PrePrints 2:e400v1. 
}
\author{
Conrad Burden
}
\examples{
cds <- makeExampleCountDataSet()
cds <- estimateSizeFactors( cds )
cds <- estimateDispersions( cds )
nbTPolyfit <- pfNbinomTest( cds, "A", "B" )
lP <- levelPValues(nbTPolyfit$pval, plot=TRUE)
pvalTab <- cbind(origPval=nbTPolyfit$pval, correctedPval=lP$pValueCorr)
cat("\n Original and corrected P-values from DESeq \n")
head(pvalTab)
}
